updated files

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: ana-pdos <ana-pdos@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/13 21:01:48 by ana-pdos          #+#    #+#             */
-/*   Updated: 2025/05/13 21:04:20 by ana-pdos         ###   ########.fr       */
+/*   Updated: 2025/05/16 14:45:09 by ana-pdos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*a;
 
 	a = (unsigned char *)s;
